backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pandas==2.2.1
numpy==1.26.4
scikit-learn
networkx
